Move geometry classes to geometry folder
