+++ trview.app/Compass.cpp
@@ -1,6 +1,6 @@
 #include "Compass.h"
 #include "Camera/ICamera.h"
-#include "MeshVertex.h"
+#include <trview.app/Geometry/MeshVertex.h>
 #include <trview.graphics/Sprite.h>
 #include <trview.graphics/IShaderStorage.h>
 #include <trview.graphics/RenderTargetStore.h>
