+++ trview.app/Geometry/IRenderable.cpp
