+++ trview.app/Geometry/IRenderable.h
