+++ trview.app/Geometry/MeshVertex.h
