+++ trview.app/Geometry/TransparentTriangle.cpp
