+++ trview.app/Geometry/TransparentTriangle.h
