+++ trview.app/Geometry/Triangle.h
