+++ trview/Entity.cpp
@@ -7,7 +7,7 @@
 #include <trview.app/ILevelTextureStorage.h>
 #include "IMeshStorage.h"
 #include <trview.app/Camera/ICamera.h>
-#include <trview.app/Mesh.h>
+#include <trview.app/Geometry/Mesh.h>
 #include <trview.app/TransparencyBuffer.h>
 
 #include <trlevel/ILevel.h>
