+++ trview/MeshStorage.h
@@ -9,7 +9,7 @@
 #include <trlevel/ILevel.h>
 
 #include "IMeshStorage.h"
-#include <trview.app/Mesh.h>
+#include <trview.app/Geometry/Mesh.h>
 #include <trview.graphics/Device.h>
 
 namespace trview
