+++ trview/StaticMesh.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 #include "StaticMesh.h"
-#include <trview.app/Mesh.h>
+#include <trview.app/Geometry/Mesh.h>
 #include <trview.app/TransparencyBuffer.h>
 
 namespace trview
